requirements.txt
streamlit>=1.30.0
pandas>=2.2.0
numpy>=1.25.0
plotly>=5.20.0
Pillow>=10.0.0
gspread>=5.11.3
oauth2client>=4.1.3
openpyxl>=3.1.2
